Move timer from Model to MainWindow
